@@ -2,11 +2,12 @@
   "name": "gradestats-app",
   "version": "1.12.2",
   "private": true,
+  "type": "module",
   "engines": {
-    "node": "16.x"
+    "node": ">=22.x"
   },
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "format": "prettier --write './**/*.{js,jsx,ts,tsx,css,scss}'",
@@ -14,55 +15,42 @@
     "lint-check": "eslint './**/*.{js,jsx,ts,tsx}'",
     "lint": "eslint './**/*.{js,jsx,ts,tsx}' --fix",
     "tsc": "tsc --noEmit",
-    "release": "semantic-release",
     "lighthouse": "lhci autorun"
   },
   "dependencies": {
-    "@octokit/rest": "^18.0.12",
-    "@reach/dialog": "^0.12.1",
-    "@sentry/browser": "^6.0.1",
-    "@sentry/node": "^6.0.1",
-    "@sentry/webpack-plugin": "^1.11.1",
-    "@types/classnames": "^2.2.11",
-    "@types/core-js": "^2.5.4",
-    "@types/gtag.js": "0.0.4",
-    "@types/next-auth": "^3.1.21",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@zeit/next-source-maps": "^0.0.3",
-    "classnames": "^2.2.6",
-    "core-js": "^3.8.3",
+    "@octokit/rest": "^21.1.1",
+    "@sentry/nextjs": "^9.5.0",
+    "@types/gtag.js": "0.0.20",
+    "@types/react-dom": "^19.0.4",
+    "classnames": "^2.5.1",
     "cssremedy": "^0.1.0-beta.2",
-    "intersection-observer": "^0.12.0",
-    "next": "^10.0.4",
-    "next-auth": "^3.1.0",
-    "react": "^17.0.1",
-    "react-alice-carousel": "^2.2.0",
-    "react-dom": "^17.0.1",
-    "sass": "^1.29.0",
-    "scroll-behavior-polyfill": "^2.0.13",
-    "swr": "^0.4.0",
-    "typescript": "^4.1.3",
-    "victory": "^35.4.4"
+    "next": "^15.2.0",
+    "next-auth": "^4.24.11",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-slick": "^0.30.3",
+    "sass": "^1.85.1",
+    "slick-carousel": "^1.8.1",
+    "swr": "^2.3.2",
+    "typescript": "^5.7.3",
+    "victory": "^37.3.6"
   },
   "devDependencies": {
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-conventional": "^11.0.0",
-    "@lhci/cli": "^0.6.1",
-    "@semantic-release/commit-analyzer": "^8.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@semantic-release/github": "^7.2.0",
-    "@semantic-release/npm": "^7.0.9",
-    "@semantic-release/release-notes-generator": "^9.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.11.0",
-    "@typescript-eslint/parser": "^4.11.0",
-    "conventional-changelog-conventionalcommits": "^4.5.0",
-    "eslint": "^7.16.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-react": "^7.19.0",
-    "husky": "^4.3.6",
-    "prettier": "^2.0.5",
-    "semantic-release": "^17.3.1"
+    "@eslint/eslintrc": "^3.3.0",
+    "@eslint/js": "^9.21.0",
+    "@lhci/cli": "^0.14.0",
+    "@next/eslint-plugin-next": "^15.2.0",
+    "@types/react": "^19.0.10",
+    "@types/react-slick": "^0.23.13",
+    "@typescript-eslint/eslint-plugin": "^8.25.0",
+    "@typescript-eslint/parser": "^8.25.0",
+    "eslint": "^9.21.0",
+    "eslint-config-prettier": "^10.0.2",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-compiler": "^19.0.0-beta-e1e972c-20250221",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "prettier": "^3.5.2",
+    "typescript-eslint": "^8.25.0"
   }
 }
