@@ -1,14 +1,14 @@
-import { FC } from 'react';
-import { useUser } from 'common/hooks/useUser';
+import type { FC } from 'react';
 import dynamic from 'next/dynamic';
-import { useRouter } from 'next/router';
+import { usePathname, useRouter } from 'next/navigation';
 import { Heading } from 'components/Typography/Heading';
 import { Text } from 'components/Typography/Text';
 import { Button } from 'components/common/Button';
 
 import styles from './tags.module.scss';
 import { useDialog } from 'common/hooks/useDialog';
-import { Tag } from 'models/Tag';
+import type { Tag } from 'models/Tag';
+import { useSession } from 'next-auth/react';
 
 const DynamicAddTagDialog = dynamic(() => import('./AddTagDialog'), { ssr: false });
 
@@ -18,13 +18,15 @@ interface Props {
 }
 
 export const Tags: FC<Props> = ({ courseCode, tags }) => {
-  const [user] = useUser();
+  const { data: session } = useSession();
   const router = useRouter();
-  const [showDialog, openDialog, closeDialog] = useDialog();
+  const path = usePathname();
+
+  const [showDialog, openDialog, closeDialog] = useDialog(false);
 
   const handleClick = () => {
-    if (!user) {
-      router.push({ pathname: '/login', query: { returnToPath: router.asPath } });
+    if (!session?.user) {
+      router.push(`/login?returnToPath=${path}`);
     } else {
       openDialog();
     }
@@ -38,7 +40,7 @@ export const Tags: FC<Props> = ({ courseCode, tags }) => {
       <Button type="button" onClick={handleClick}>
         Legg til tags for søking
       </Button>
-      {!user && (
+      {!session?.user && (
         <Text className={styles.loginText}>
           <small>Du må være logget inn for å kunne legge til tags.</small>
         </Text>
