@@ -1,27 +1,29 @@
-import { RefObject, useEffect, useRef, useState } from 'react';
+'use client';
+import type { RefObject } from 'react';
+import { useEffect, useLayoutEffect, useRef, useState } from 'react';
 
 export type Ref = RefObject<HTMLDivElement>;
 export type Entry = IntersectionObserverEntry;
 
-export const useIntersection = (callback?: () => void, options?: IntersectionObserverInit): Ref => {
+export const useIntersection = (
+  callback?: () => void,
+  options?: IntersectionObserverInit
+): RefObject<HTMLDivElement | null> => {
   const targetRef = useRef<HTMLDivElement>(null);
   const [observerEntry, setEntry] = useState<Entry | null>(null);
   const [count, setCount] = useState(0);
 
-  useEffect(() => {
-    if (count !== 0) {
-      const observer = new IntersectionObserver((entries) => {
-        const [firstEntry] = entries;
-        setEntry(firstEntry);
-      }, options);
-      if (targetRef.current) {
-        observer.observe(targetRef.current);
-      }
-      return () => observer.disconnect();
+  useLayoutEffect(() => {
+    const observer = new IntersectionObserver((entries) => {
+      const [firstEntry] = entries;
+      setEntry(firstEntry);
+    }, options);
+    if (targetRef.current) {
+      observer.observe(targetRef.current);
     }
     setCount((oldCount) => oldCount + 1);
-    return;
-  }, [targetRef.current]);
+    return () => observer.disconnect();
+  }, []);
 
   useEffect(() => {
     if (observerEntry?.isIntersecting && count !== 0) {
