@@ -1,7 +1,8 @@
 import { graphTheme } from 'common/utils/chart';
 import { getColorForGradeLetter } from 'common/utils/grades';
-import { Grade } from 'models/Grade';
-import { FC, memo } from 'react';
+import type { Grade } from 'models/Grade';
+import type { FC } from 'react';
+import { memo } from 'react';
 import { VictoryBar, VictoryChart, VictoryAxis } from 'victory';
 
 interface Props {
