@@ -1,18 +1,18 @@
+'use client';
 import dynamic from 'next/dynamic';
-import { FC, useState } from 'react';
-import Carousel from 'react-alice-carousel';
+import type { FC } from 'react';
+import { useRef, useState } from 'react';
+import Slider from 'react-slick';
 import cx from 'classnames';
 
 import styles from './course-charts.module.scss';
-
+import './dots.css';
 import { Button } from 'components/common/Button';
-import { Grade } from 'models/Grade';
+import type { Grade } from 'models/Grade';
 import GradedGraph from './GradedGraph';
 import UngradedGraph from './UngradedGraph';
-import { EventObject } from 'react-alice-carousel/lib/types';
 import { Heading } from 'components/Typography/Heading';
 import { BasicCard } from 'components/Card/BasicCard';
-import { CarouselDot } from './CarouselDot';
 
 const DynamicAverageChart = dynamic(() => import('./AverageChart'), { ssr: false });
 const DynamicFailedChart = dynamic(() => import('./FailedChart'), { ssr: false });
@@ -44,19 +44,26 @@ const TAB_TITLE: Record<number, string> = {
 
 export const CourseCharts: FC<Props> = ({ className, grades, currentGrade }) => {
   const [tab, setTab] = useState<Tab>('BAR');
+
+  const settings = {
+    dots: true,
+    infinite: true,
+    speed: 500,
+    slidesToShow: 1,
+    slidesToScroll: 1,
+    beforeChange: (current: number, next: number) => setTab(INDEX_TAB[next]),
+  };
+
+  const sliderRef = useRef<Slider>(null);
+  const next = () => {
+    sliderRef.current?.slickNext();
+  };
+
   const courseHasGrades = grades.length > 0;
 
   if (!courseHasGrades) {
     return null;
   }
-
-  const handleSlideChange = (event: EventObject) => {
-    setTab(INDEX_TAB[event.item]);
-  };
-
-  const handleNextSlideClick = () => {
-    setTab(INDEX_TAB[(TAB_INDEX[tab] + 1) % 3]);
-  };
 
   return (
     <div className={cx(styles.victoryContainer, className)}>
@@ -64,36 +71,34 @@ export const CourseCharts: FC<Props> = ({ className, grades, currentGrade }) => 
         <Heading as="h2" size="h4">
           {TAB_TITLE[TAB_INDEX[tab]]}
         </Heading>
-        <Button variant="link" onClick={handleNextSlideClick}>
+        <Button variant="link" onClick={next}>
           <Heading as="p" size="h4">
             {TAB_TITLE[(TAB_INDEX[tab] + 1) % 3] + ` ->`}
           </Heading>
         </Button>
       </header>
-      <Carousel
-        mouseTracking
-        activeIndex={TAB_INDEX[tab]}
-        onSlideChanged={handleSlideChange}
-        disableButtonsControls
-        renderDotsItem={CarouselDot}
-        items={[
-          currentGrade.passed === 0 ? (
-            <BasicCard key="graded">
-              <GradedGraph grade={currentGrade} />
-            </BasicCard>
-          ) : (
-            <BasicCard key="ungraded">
-              <UngradedGraph grade={currentGrade} />
-            </BasicCard>
-          ),
-          <BasicCard key="averages">
-            <DynamicAverageChart grades={grades} currentSemesterCode={currentGrade.semester_code} />
-          </BasicCard>,
-          <BasicCard key="failed">
-            <DynamicFailedChart grades={grades} />
-          </BasicCard>,
-        ]}
-      />
+      <Slider
+        ref={(slider) => {
+          sliderRef.current = slider;
+        }}
+        {...settings}
+      >
+        {currentGrade.passed === 0 ? (
+          <BasicCard key="graded">
+            <GradedGraph grade={currentGrade} />
+          </BasicCard>
+        ) : (
+          <BasicCard key="ungraded">
+            <UngradedGraph grade={currentGrade} />
+          </BasicCard>
+        )}
+        <BasicCard key="averages">
+          <DynamicAverageChart grades={grades} currentSemesterCode={currentGrade.semester_code} />
+        </BasicCard>
+        <BasicCard key="failed">
+          <DynamicFailedChart grades={grades} />
+        </BasicCard>
+      </Slider>
     </div>
   );
 };
