@@ -1,7 +1,8 @@
+'use client';
 import { useDialog } from 'common/hooks/useDialog';
 import { Button } from 'components/common/Button';
 import dynamic from 'next/dynamic';
-import { FC } from 'react';
+import type { FC } from 'react';
 
 const DynamicReportDialog = dynamic(() => import('./ReportDialog'), { ssr: false });
 
@@ -9,6 +10,7 @@ interface Props {
   className?: string;
   text?: string;
   courseCode?: string;
+  children?: React.ReactNode;
 }
 
 export const ReportDialogButton: FC<Props> = ({ className, courseCode, text, children }) => {
