@@ -1,4 +1,5 @@
-import { forwardRef, HTMLProps } from 'react';
+import type { HTMLProps } from 'react';
+import { forwardRef } from 'react';
 import cx from 'classnames';
 
 import styles from './select.module.scss';
