@@ -2,9 +2,8 @@ import { mapGradeAverageToLetter } from 'common/utils/grades';
 import { formatPercentage } from 'common/utils/math';
 import { LinkCard } from 'components/Card/LinkCard';
 import { Text } from 'components/Typography/Text';
-import { Course } from 'models/Course';
-import Link from 'next/link';
-import { FC } from 'react';
+import type { Course } from 'models/Course';
+import type { FC } from 'react';
 import cx from 'classnames';
 
 import styles from './course-item.module.scss';
@@ -20,23 +19,19 @@ export const CourseItem: FC<Props> = ({ course }) => {
 
   return (
     <li className={styles.item}>
-      <Link href="/course/[courseCode]" as={`/course/${course.code}`}>
-        <LinkCard className={styles.content}>
-          <Text>{course.code}</Text>
-          <Text>{course.norwegian_name}</Text>
-          <Text>
-            {has_grades ? (
-              showGradeLetter ? (
-                <Text size="h4">{mapGradeAverageToLetter(averageGrade)}</Text>
-              ) : (
-                <Text>{averageGrade > 0 ? formatPercentage(averageGrade) : '-'}</Text>
-              )
-            ) : (
-              '-'
-            )}
-          </Text>
-        </LinkCard>
-      </Link>
+      <LinkCard href={`/course/${course.code}`} className={styles.content}>
+        <Text>{course.code}</Text>
+        <Text>{course.norwegian_name}</Text>
+        {has_grades ? (
+          showGradeLetter ? (
+            <Text size="h4">{mapGradeAverageToLetter(averageGrade)}</Text>
+          ) : (
+            <Text>{averageGrade > 0 ? formatPercentage(averageGrade) : '-'}</Text>
+          )
+        ) : (
+          <Text>-</Text>
+        )}
+      </LinkCard>
     </li>
   );
 };
