@@ -1,11 +1,11 @@
 import { SearchInput } from 'components/forms/SearchInput';
 import { InifiniteLoading } from 'components/Loading/InfiniteLoading';
 import { Text } from 'components/Typography/Text';
-import { Course, CourseSort } from 'models/Course';
-import { Department } from 'models/Department';
-import { Faculty } from 'models/Faculty';
-import Head from 'next/head';
-import { ChangeEvent, FC, MutableRefObject, useState } from 'react';
+import type { Course, CourseSort } from 'models/Course';
+import type { Department } from 'models/Department';
+import type { Faculty } from 'models/Faculty';
+import type { ChangeEvent, FC, RefObject } from 'react';
+import { useState } from 'react';
 import { CourseItem, CourseListHeader } from './CourseItem';
 
 import styles from './course-list-view.module.scss';
@@ -14,7 +14,7 @@ import { CourseFilters } from './CourseFilters';
 import { FilterIcon } from 'components/Graphics/Icons/Filter';
 
 interface Props {
-  searchBarRef: MutableRefObject<HTMLInputElement | null>;
+  searchBarRef: RefObject<HTMLInputElement | null>;
   onSearchChange: (query: string) => void;
   onOrderingChange: (sortOrder: CourseSort) => void;
   currentOrdering: CourseSort;
@@ -58,63 +58,55 @@ export const CourseListView: FC<Props> = ({
   const toggleShowMenu = () => setShowFilters((current) => !current);
 
   return (
-    <>
-      <Head>
-        <title>grades.no - søk</title> 
-        <meta property="og:title" content="grades.no - søk" />
-        <meta name="description" content="Søk i emner ved NTNU" />
-        <meta property="og:description" content="Søk i emner ved NTNU" />
-      </Head>
-      <section className={styles.container}>
-        <div className={styles.searchContainer}>
-          <label className={styles.searchLabel} htmlFor="search">
-            Søk i emner
-          </label>
-          <SearchInput
-            id="search"
-            className={styles.searchBar}
-            ref={searchBarRef}
-            placeholder="Søk i emner..."
-            aria-label="Søk i emner"
-            onChange={handleSearchChange}
-            value={query}
-          />
-          <Button
-            className={styles.toggleFiltersButton}
-            active={showFilters}
-            invertedActive
-            onClick={toggleShowMenu}
-            aria-expanded={showFilters}
-            aria-label="Vis søkefiltere"
-          >
-            <FilterIcon className={styles.filtersIcon} />
-          </Button>
-        </div>
-        {showFilters ? (
-          <CourseFilters
-            onOrderingChange={onOrderingChange}
-            currentOrdering={currentOrdering}
-            onFacultyFilterChange={onFacultyFilterChange}
-            onDepartmentFilterChange={onDepartmentFilterChange}
-            departments={departments}
-            currentDepartmentId={currentDepartmentId}
-            faculties={faculties}
-            currentFacultyId={currentFacultyId}
-          />
-        ) : null}
-        {!isLoading && !courses.length && <Text>Ingen resultater</Text>}
-        {courses.length ? (
-          <>
-            <CourseListHeader />
-            <ol className={styles.courseList}>
-              {courses.map((course) => (
-                <CourseItem key={course.code} course={course} />
-              ))}
-            </ol>
-          </>
-        ) : null}
-        <InifiniteLoading isLoading={isLoading} triggerNextPage={nextPage} />
-      </section>
-    </>
+    <section className={styles.container}>
+      <div className={styles.searchContainer}>
+        <label className={styles.searchLabel} htmlFor="search">
+          Søk i emner
+        </label>
+        <SearchInput
+          id="search"
+          className={styles.searchBar}
+          ref={searchBarRef}
+          placeholder="Søk i emner..."
+          aria-label="Søk i emner"
+          onChange={handleSearchChange}
+          value={query}
+        />
+        <Button
+          className={styles.toggleFiltersButton}
+          active={showFilters}
+          invertedActive
+          onClick={toggleShowMenu}
+          aria-expanded={showFilters}
+          aria-label="Vis søkefiltere"
+        >
+          <FilterIcon className={styles.filtersIcon} />
+        </Button>
+      </div>
+      {showFilters ? (
+        <CourseFilters
+          onOrderingChange={onOrderingChange}
+          currentOrdering={currentOrdering}
+          onFacultyFilterChange={onFacultyFilterChange}
+          onDepartmentFilterChange={onDepartmentFilterChange}
+          departments={departments}
+          currentDepartmentId={currentDepartmentId}
+          faculties={faculties}
+          currentFacultyId={currentFacultyId}
+        />
+      ) : null}
+      {!isLoading && !courses.length && <Text>Ingen resultater</Text>}
+      {courses.length ? (
+        <>
+          <CourseListHeader />
+          <ol className={styles.courseList}>
+            {courses.map((course) => (
+              <CourseItem key={course.code} course={course} />
+            ))}
+          </ol>
+        </>
+      ) : null}
+      <InifiniteLoading isLoading={isLoading} triggerNextPage={nextPage} />
+    </section>
   );
 };
