@@ -1,6 +1,5 @@
 import { mapGradeAverageToLetter } from 'common/utils/grades';
-import Link from 'next/link';
-import { FC } from 'react';
+import type { FC } from 'react';
 import { Heading } from 'components/Typography/Heading';
 import { Text } from 'components/Typography/Text';
 import cx from 'classnames';
@@ -8,7 +7,7 @@ import cx from 'classnames';
 import styles from './course-card.module.scss';
 import { LinkCard } from 'components/Card/LinkCard';
 import { formatPercentage } from 'common/utils/math';
-import { Course } from 'models/Course';
+import type { Course } from 'models/Course';
 
 export const GRADE_COLORS: Record<string, string> = {
   A: styles.gradeA,
@@ -34,16 +33,14 @@ export const CourseCard: FC<Props> = ({ className, code, name, course }) => {
 
   const gradeLetter = averageGrade ? mapGradeAverageToLetter(averageGrade) : undefined;
   return (
-    <Link href="/course/[courseCode]" as={`/course/${code}`}>
-      <LinkCard className={cx(className, styles.card)}>
-        <Heading className={styles.code} as="h3">
-          {code}
-        </Heading>
-        <Text className={styles.text}>{name}</Text>
-        <Text className={cx(styles.grade, GRADE_COLORS[gradeLetter ?? 'PASSED'])}>
-          {showGradeLetter ? gradeLetter || '-' : averageGrade ? formatPercentage(averageGrade) : '-'}
-        </Text>
-      </LinkCard>
-    </Link>
+    <LinkCard href={`/course/${code}`} className={cx(className, styles.card)}>
+      <Heading className={styles.code} as="h3">
+        {code}
+      </Heading>
+      <Text className={styles.text}>{name}</Text>
+      <Text className={cx(styles.grade, GRADE_COLORS[gradeLetter ?? 'PASSED'])}>
+        {showGradeLetter ? gradeLetter || '-' : averageGrade ? formatPercentage(averageGrade) : '-'}
+      </Text>
+    </LinkCard>
   );
 };
