@@ -1,13 +1,12 @@
-import { ChangeEvent, FC } from 'react';
+import type { ChangeEvent, FC } from 'react';
 import useSWR from 'swr';
 import cx from 'classnames';
 
 import { requestsWithAuth } from 'common/requests';
 import { getUserDetailApiUrl } from 'common/urls';
 import { formatDateString } from 'common/utils/date';
-import { LinkCard } from 'components/Card/LinkCard';
 import { Text } from 'components/Typography/Text';
-import { GradesUser } from 'models/User';
+import type { GradesUser } from 'models/User';
 
 import styles from './user-item.module.scss';
 import { BasicCard } from 'components/Card/BasicCard';
@@ -29,7 +28,7 @@ const TEXT = {
 export const UserItem: FC<Props> = ({ user: initialUser }) => {
   const userUrl = getUserDetailApiUrl(initialUser.username);
   const { data, mutate } = useSWR<GradesUser>(userUrl, requestsWithAuth.get, {
-    initialData: initialUser,
+    fallbackData: initialUser,
   });
   const user = data as GradesUser;
 
